Fix path to pull helm charts
Helm charts are published under `nginx-azure-lb`
alongside docker images for all of our services.

In addition to the above, charts are published
keeping in mind that each feature branch iteration
can produce a helm chart. While doing a release,
we need to pull in the main chart and then re-tag
it to push to docker hub.