Change extension to ClusterScope
When trying to install the extension in a
namespaced mode (which is the default on the
current offer/plan), the AKS extension machinery
creates a k8s service account (different from the
NLK service account) which is restricted to the
extension namespace. Due to the aforementioned,
the service account cannot read Cluster Roles that
we create for NLK causing the extension to fail.

Switching over to the cluster level extension will
help here as when the AKS extension gets installed
cluster-wide, the undelying service account has
all the necessary permissions for the extension to
be installed.