Fix chart,app version for AKS marketplace
While nothing breaks if the versions are not
aligned inside the AKS marketplace application
bundle, the side effect is that when a user review
the helm releases on an AKS cluster, they see the
hardcoded app and chart version which is
misleading.

This commit fixes the chart and app version inside
the Helm chart so that the user sees correct
information while interacting with their AKS
cluster.