Use upstream SAST template
We were using the deprecated internal SAST
template which was retired a while back:
https://gitlab.com/f5/nginx/tools/devops-utils/-/merge_requests/204.

There was a change made to the upstream template
that is included in the internal template which
breaks the internal template as there were some
jobs that were removed from the upstream.

This commit moves us over to use the official
upstream template.